+++ pruebas/01-reading-list/lmh2022/src/components/ReadBooks/ReadBooks.css
@@ -8,7 +8,7 @@
 }
 
 
-.book {
+.read-book {
     background-size: contain;
     background-repeat: no-repeat;
     background-position: center;
